requirements.txt
ultralytics>=8.3
labelme>=5.4
dvc[s3]>=3.50
mlflow>=2.12
pyyaml>=6.0
Pillow>=10.0
pytest>=8.0
